Defer values to outer scope if they do not exist
If the property does not exist, we should not replace it, as it could be
picked up by an outer render.

For example, in a loop, the value may not be on the loop object, but in
the original object containing the loop property. This will be handled
after the loop render completes.